move legend style step to own module:
- un-exposed legend.lines, legend.bars, legend.points,
  legend.boxes and legend.pie
