+++ src/components/legend/index.js
@@ -29,207 +29,6 @@ var constants = require('./constants');
 legend.layoutAttributes = require('./attributes');
 
 legend.supplyLayoutDefaults = require('./defaults');
-// -----------------------------------------------------
-// styling functions for traces in legends.
-// same functions for styling traces in the popovers
-// -----------------------------------------------------
-
-legend.lines = function(d) {
-    var trace = d[0].trace,
-        showFill = trace.visible && trace.fill && trace.fill!=='none',
-        showLine = subTypes.hasLines(trace);
-
-    var fill = d3.select(this).select('.legendfill').selectAll('path')
-        .data(showFill ? [d] : []);
-    fill.enter().append('path').classed('js-fill',true);
-    fill.exit().remove();
-    fill.attr('d', 'M5,0h30v6h-30z')
-        .call(Drawing.fillGroupStyle);
-
-    var line = d3.select(this).select('.legendlines').selectAll('path')
-        .data(showLine ? [d] : []);
-    line.enter().append('path').classed('js-line',true)
-        .attr('d', 'M5,0h30');
-    line.exit().remove();
-    line.call(Drawing.lineGroupStyle);
-};
-
-legend.points = function(d) {
-    var d0 = d[0],
-        trace = d0.trace,
-        showMarkers = subTypes.hasMarkers(trace),
-        showText = subTypes.hasText(trace),
-        showLines = subTypes.hasLines(trace);
-
-    var dMod, tMod;
-
-    // 'scatter3d' and 'scattergeo' don't use gd.calcdata yet;
-    // use d0.trace to infer arrayOk attributes
-
-    function boundVal(attrIn, arrayToValFn, bounds) {
-        var valIn = Lib.nestedProperty(trace, attrIn).get(),
-            valToBound = (Array.isArray(valIn) && arrayToValFn) ?
-                arrayToValFn(valIn) : valIn;
-
-        if(bounds) {
-            if(valToBound < bounds[0]) return bounds[0];
-            else if(valToBound > bounds[1]) return bounds[1];
-        }
-        return valToBound;
-    }
-
-    function pickFirst(array) { return array[0]; }
-
-    // constrain text, markers, etc so they'll fit on the legend
-    if(showMarkers || showText || showLines) {
-        var dEdit = {},
-            tEdit = {};
-
-        if(showMarkers) {
-            dEdit.mc = boundVal('marker.color', pickFirst);
-            dEdit.mo = boundVal('marker.opacity', Lib.mean, [0.2, 1]);
-            dEdit.ms = boundVal('marker.size', Lib.mean, [2, 16]);
-            dEdit.mlc = boundVal('marker.line.color', pickFirst);
-            dEdit.mlw = boundVal('marker.line.width', Lib.mean, [0, 5]);
-            tEdit.marker = {
-                sizeref: 1,
-                sizemin: 1,
-                sizemode: 'diameter'
-            };
-        }
-
-        if(showLines) {
-            tEdit.line = {
-                width: boundVal('line.width', pickFirst, [0, 10])
-            };
-        }
-
-        if(showText) {
-            dEdit.tx = 'Aa';
-            dEdit.tp = boundVal('textposition', pickFirst);
-            dEdit.ts = 10;
-            dEdit.tc = boundVal('textfont.color', pickFirst);
-            dEdit.tf = boundVal('textfont.family', pickFirst);
-        }
-
-        dMod = [Lib.minExtend(d0, dEdit)];
-        tMod = Lib.minExtend(trace, tEdit);
-    }
-
-    var ptgroup = d3.select(this).select('g.legendpoints');
-
-    var pts = ptgroup.selectAll('path.scatterpts')
-        .data(showMarkers ? dMod : []);
-    pts.enter().append('path').classed('scatterpts', true)
-        .attr('transform', 'translate(20,0)');
-    pts.exit().remove();
-    pts.call(Drawing.pointStyle, tMod);
-
-    // 'mrc' is set in pointStyle and used in textPointStyle:
-    // constrain it here
-    if(showMarkers) dMod[0].mrc = 3;
-
-    var txt = ptgroup.selectAll('g.pointtext')
-        .data(showText ? dMod : []);
-    txt.enter()
-        .append('g').classed('pointtext',true)
-            .append('text').attr('transform', 'translate(20,0)');
-    txt.exit().remove();
-    txt.selectAll('text').call(Drawing.textPointStyle, tMod);
-
-};
-
-legend.bars = function(d) {
-    var trace = d[0].trace,
-        marker = trace.marker||{},
-        markerLine = marker.line||{},
-        barpath = d3.select(this).select('g.legendpoints')
-            .selectAll('path.legendbar')
-            .data(Plots.traceIs(trace, 'bar') ? [d] : []);
-    barpath.enter().append('path').classed('legendbar',true)
-        .attr('d','M6,6H-6V-6H6Z')
-        .attr('transform','translate(20,0)');
-    barpath.exit().remove();
-    barpath.each(function(d) {
-        var w = (d.mlw+1 || markerLine.width+1) - 1,
-            p = d3.select(this);
-        p.style('stroke-width',w+'px')
-            .call(Color.fill, d.mc || marker.color);
-        if(w) {
-            p.call(Color.stroke, d.mlc || markerLine.color);
-        }
-    });
-};
-
-legend.boxes = function(d) {
-    var trace = d[0].trace,
-        pts = d3.select(this).select('g.legendpoints')
-            .selectAll('path.legendbox')
-            .data(Plots.traceIs(trace, 'box') && trace.visible ? [d] : []);
-    pts.enter().append('path').classed('legendbox', true)
-        // if we want the median bar, prepend M6,0H-6
-        .attr('d', 'M6,6H-6V-6H6Z')
-        .attr('transform', 'translate(20,0)');
-    pts.exit().remove();
-    pts.each(function(d) {
-        var w = (d.lw+1 || trace.line.width+1) - 1,
-            p = d3.select(this);
-        p.style('stroke-width', w+'px')
-            .call(Color.fill, d.fc || trace.fillcolor);
-        if(w) {
-            p.call(Color.stroke, d.lc || trace.line.color);
-        }
-    });
-};
-
-legend.pie = function(d) {
-    var trace = d[0].trace,
-        pts = d3.select(this).select('g.legendpoints')
-            .selectAll('path.legendpie')
-            .data(Plots.traceIs(trace, 'pie') && trace.visible ? [d] : []);
-    pts.enter().append('path').classed('legendpie', true)
-        .attr('d', 'M6,6H-6V-6H6Z')
-        .attr('transform', 'translate(20,0)');
-    pts.exit().remove();
-
-    if(pts.size()) pts.call(styleOne, d[0], trace);
-};
-
-legend.style = function(s) {
-    s.each(function(d) {
-        var traceGroup = d3.select(this);
-
-        var fill = traceGroup
-            .selectAll('g.legendfill')
-                .data([d]);
-        fill.enter().append('g')
-            .classed('legendfill',true);
-
-        var line = traceGroup
-            .selectAll('g.legendlines')
-                .data([d]);
-        line.enter().append('g')
-            .classed('legendlines',true);
-
-        var symbol = traceGroup
-            .selectAll('g.legendsymbols')
-                .data([d]);
-        symbol.enter().append('g')
-            .classed('legendsymbols',true);
-        symbol.style('opacity', d[0].trace.opacity);
-
-        symbol.selectAll('g.legendpoints')
-            .data([d])
-          .enter().append('g')
-            .classed('legendpoints',true);
-    })
-    .each(legend.bars)
-    .each(legend.boxes)
-    .each(legend.pie)
-    .each(legend.lines)
-    .each(legend.points);
-};
-
 legend.texts = function(context, td, d, i, traces) {
     var fullLayout = td._fullLayout,
         trace = d[0].trace,
@@ -724,6 +523,7 @@ legend.repositionLegend = function(td, traces) {
     lx = Math.round(lx);
     ly = Math.round(ly);
 
+legend.style = require('./style');
     // lastly check if the margin auto-expand has changed
     Plots.autoMargin(td,'legend',{
         x: opts.x,
